requirements.txt
nltk==3.8.1
openai==0.27.2
tqdm==4.65.0
pdfminer.six==20221105
